+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 2: JavaScript Language Fundamentals
     - Variables - var, let & const
     - Data Types in JavaScript
+    - Type Conversion
 - An example with var, let and const for Variables - var, let & const was added.
 - An example with data types for Data Types in JavaScript was added.
+- An example with type conversion for Type Conversion was added.
 
 ## July 30, 2020
 - Modern JavaScript From The Beginning (Udemy)